app/src/requirements.txt
flask==2.0.1
uWSGI==2.0.19.1
pandas==1.3.4
textfsm==1.1.2
ruamel.yaml==0.17.17
jinja2==3.0.2
mysqlclient==2.1.0
sqlalchemy
flask-sqlalchemy
pymysql
mysql